removed session stylesheets
Made columns work right with foundation. On large
screens the poster is to the left of the information.

On small screens the poster is below the title and
ratings.

Added styles for recommendation

* Changed the method that creates my own personal
  credentials I log in with

* Changed the user_params method to accept input
  so I could send in my own credentials

* Genres now samples a set of genres between 1 and
  3

renamed the recommendation stylesheet to css.scss

Styled recommendation to have the plot under the
title

replaced web-console with better_errors

converted recommendation into a partial so it
can be used on other sections of the app