added private to the omdb_service model
Fixed the url on the get request to work by moving the
“?”

the media method now accepts a media type

wrote tests for it

Added the service to the recommendation model

Added the find_media method that will find the media
for the user with the default of movie. It also takes
the year as a parameter if needed.

Added a build recommendation model that returns
a recommendation object

renamed the find media method to match new functionality
of creating a movie if it’s not in the database

refactored the build recommendation into it’s own class,
recommendation parser

had to add the silly data that I’m not going to use
anyway

yay tests that test that the test is testing things